requirements.txt
argparse; python_version < "3.11"
faster-whisper==0.10.1
llama-cpp-python==0.2.84
numpy<2.0.0
soundfile==0.12.1
TTS==0.22.0
av==12.1.0
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9
